Connect to MQTT on start
