--- website/index.js
@@ -1,12 +1,0 @@
-const express = require("express");
-
-const app = express();
-const port = 3000;
-
-app.use(express.static("public"));
-
-app.get("/", (req, res) => {
-  res.send(__dirname + "/public/index.html");
-});
-
-app.listen(port, () => console.log(`Example app listening on port ${port}!`));
