+++ website/package.json
@@ -9,7 +9,10 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "body-parser": "^1.19.0",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "express-reload": "^1.2.0"
+    "express-reload": "^1.2.0",
+    "mqtt": "^3.0.0"
   }
 }
